💄 style(index): update help information comment
- Improve clarity of comment regarding help information
- Remove the "NOTE:" prefix for a cleaner look

♻️ refactor(deploy): add check for file existence before copy

- Ensure backup process works only if the tfStateFile exists
- Prevent potential errors when trying to copy a non-existent file

+++ src/services/deploy.ts
@@ -163,8 +163,10 @@ export class Deploy {
               `terraform.tfstate.${Date.now()}.backup`,
             )
             
-            // fs.renameSync(tfStateFile, backupStateFile)
-            fs.copyFileSync(tfStateFile, backupStateFile)
+            if (fs.existsSync(tfStateFile)) {
+              // fs.renameSync(tfStateFile, backupStateFile)
+              fs.copyFileSync(tfStateFile, backupStateFile)
+            }
 
             // Import existing VPC and IGW
             this.runImport(`aws_vpc.VPC "${this.enVars.VPC_ID}"`)
